Cyberchallenge/ctf_HSM+SQLI/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
pycryptodome==3.18.0